extract cpu probing into bitnet-cpu-probe microcrate
