+++ crates/bitnet-device-probe/src/lib.rs
@@ -4,6 +4,7 @@
 //! extracted from `bitnet-kernels` for use by the broader workspace.
 
 pub use bitnet_common::kernel_registry::SimdLevel;
+pub use bitnet_cpu_probe::{CpuCapabilities, detect_simd_level, probe_cpu, simd_level_rank};
 
 #[cfg(feature = "opencl")]
 pub mod opencl;
@@ -22,61 +23,6 @@ pub use wgpu_probe::{
 };
 
 // ── CPU capabilities ─────────────────────────────────────────────────────────
-
-/// CPU capabilities detected at runtime.
-///
-/// Obtained by calling [`probe_cpu`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CpuCapabilities {
-    /// Number of logical CPU cores available to the process (always ≥ 1).
-    pub core_count: usize,
-    /// AVX2 SIMD extension available on this CPU (`x86_64` only).
-    pub has_avx2: bool,
-    /// AVX-512 SIMD extension available on this CPU (`x86_64` only).
-    pub has_avx512: bool,
-    /// NEON SIMD extension available (always `true` on `AArch64`, `false` elsewhere).
-    pub has_neon: bool,
-}
-
-/// Probe the current CPU and return its capabilities.
-///
-/// `core_count` is derived from [`std::thread::available_parallelism`] and is
-/// guaranteed to be ≥ 1. SIMD flags are detected via `is_x86_feature_detected!`
-/// (`x86_64`) or compile-time cfg (`aarch64`).
-///
-/// # Examples
-///
-/// ```
-/// use bitnet_device_probe::probe_cpu;
-///
-/// let caps = probe_cpu();
-/// // At least one logical core is always present.
-/// assert!(caps.core_count >= 1);
-///
-/// // NEON and AVX flags are mutually exclusive across architectures.
-/// assert!(!(caps.has_avx2 && caps.has_neon));
-/// assert!(!(caps.has_avx512 && caps.has_neon));
-///
-/// println!(
-///     "cores={} avx2={} avx512={} neon={}",
-///     caps.core_count, caps.has_avx2, caps.has_avx512, caps.has_neon
-/// );
-/// ```
-pub fn probe_cpu() -> CpuCapabilities {
-    let core_count = std::thread::available_parallelism().map(std::num::NonZero::get).unwrap_or(1);
-
-    #[cfg(target_arch = "x86_64")]
-    let (has_avx2, has_avx512, has_neon) =
-        (is_x86_feature_detected!("avx2"), is_x86_feature_detected!("avx512f"), false);
-
-    #[cfg(target_arch = "aarch64")]
-    let (has_avx2, has_avx512, has_neon) = (false, false, true);
-
-    #[cfg(not(any(target_arch = "x86_64", target_arch = "aarch64")))]
-    let (has_avx2, has_avx512, has_neon) = (false, false, false);
-
-    CpuCapabilities { core_count, has_avx2, has_avx512, has_neon }
-}
 
 // ── GPU capabilities ─────────────────────────────────────────────────────────
 
@@ -361,44 +307,6 @@ pub const fn vulkan_available_runtime() -> bool {
     false
 }
 
-/// Detect the best SIMD instruction-set level available at runtime.
-///
-/// Detection order: AVX-512 > AVX2 > SSE4.2 (`x86_64`); NEON (`AArch64`);
-/// scalar fallback on all other targets.
-///
-/// # Examples
-///
-/// ```
-/// use bitnet_device_probe::detect_simd_level;
-///
-/// let level = detect_simd_level();
-/// println!("SIMD level: {level:?}");
-/// // level is one of: Scalar, Sse42, Avx2, Avx512, Neon
-/// ```
-#[allow(clippy::missing_const_for_fn)] // not const on x86_64 (runtime CPUID)
-pub fn detect_simd_level() -> SimdLevel {
-    #[cfg(target_arch = "x86_64")]
-    {
-        if is_x86_feature_detected!("avx512f") {
-            SimdLevel::Avx512
-        } else if is_x86_feature_detected!("avx2") {
-            SimdLevel::Avx2
-        } else if is_x86_feature_detected!("sse4.2") {
-            SimdLevel::Sse42
-        } else {
-            SimdLevel::Scalar
-        }
-    }
-    #[cfg(target_arch = "aarch64")]
-    {
-        SimdLevel::Neon
-    }
-    #[cfg(not(any(target_arch = "x86_64", target_arch = "aarch64")))]
-    {
-        SimdLevel::Scalar
-    }
-}
-
 /// Snapshot of compile-time and runtime device capabilities.
 ///
 /// Build with [`DeviceCapabilities::detect`] to capture the current machine's
@@ -519,47 +427,6 @@ pub fn probe_device() -> DeviceProbe {
         rocm_available: rocm_available_runtime(),
         npu_available: probe_npu().available,
         oneapi_available: oneapi_available_runtime(),
-    }
-}
-
-/// Return a numeric rank for a [`SimdLevel`] so callers can compare levels.
-///
-/// Higher values represent wider/more capable SIMD.
-///
-/// | Level | Rank |
-/// |-------|------|
-/// | `Scalar` | 0 |
-/// | `Sse42`  | 1 |
-/// | `Avx2`   | 2 |
-/// | `Avx512` | 3 |
-/// | `Neon`   | 4 |
-///
-/// Note: NEON and SSE/AVX are mutually exclusive instruction sets; the rank is
-/// only meaningful when comparing levels on the same architecture.
-///
-/// # Examples
-///
-/// ```
-/// use bitnet_device_probe::{simd_level_rank, SimdLevel};
-///
-/// // Wider SIMD has a strictly higher rank.
-/// assert!(simd_level_rank(&SimdLevel::Avx512) > simd_level_rank(&SimdLevel::Avx2));
-/// assert!(simd_level_rank(&SimdLevel::Avx2)   > simd_level_rank(&SimdLevel::Sse42));
-/// assert!(simd_level_rank(&SimdLevel::Sse42)  > simd_level_rank(&SimdLevel::Scalar));
-///
-/// // Scalar is always the lowest rank.
-/// assert_eq!(simd_level_rank(&SimdLevel::Scalar), 0);
-/// ```
-pub const fn simd_level_rank(level: &SimdLevel) -> u32 {
-    match level {
-        SimdLevel::Scalar => 0,
-        SimdLevel::Sse42 => 1,
-        SimdLevel::Avx2 => 2,
-        SimdLevel::Avx512 => 3,
-        SimdLevel::Neon => 4,
-        // Future variants default to a high rank so they don't compare lower than
-        // known levels when SimdLevel gains new entries.
-        _ => u32::MAX,
     }
 }
